Support DWRF stripes above MAX_INT rows in reader
OrcReader fails when a stripe has more than INT32_MAX rows.
Both ORC/DWRF format uses long for stripe number of rows.
We ran into this edge case on a file with only string as a column
and file is produced by a different writer compatible with
DWRF spec.

+++ presto-orc/src/main/java/com/facebook/presto/orc/AbstractOrcRecordReader.java
@@ -689,7 +689,7 @@ abstract class AbstractOrcRecordReader<T extends StreamReader>
         }
     }
 
-    private void validateWriteStripe(int rowCount)
+    private void validateWriteStripe(long rowCount)
     {
         if (writeChecksumBuilder.isPresent()) {
             writeChecksumBuilder.get().addStripe(rowCount);
